feat(desktop): pricing breakdown moves into a dedicated modal
PR10 of pre-1.0 ux overhaul. Footer pricing pill now opens a Dialog
(same modal pattern as Settings) instead of a tiny popover.

apps/desktop/src/components/PricingDialog.tsx (new):
- size="lg" Dialog with title "pricing"
- two stat cards at top: current session cost + workspace total
- summarizer cost badge shown when > 0
- by-provider section: progress bars with cap %
- by-model table: provider · model | in tokens | out tokens | cost
- turns log: existing recent/expensive sort, scrollable

apps/desktop/src/components/TelemetryPill.tsx:
- now a thin button — same cost glance, on click opens PricingDialog
- removed the inline popover, sort state, and breakdown rendering
- ~210 lines → 35 lines

Verification: desktop typecheck green; 172/172 tests pass.
